+++ integration/stencil-toolkit-nx/libs/react-lib/src/lib/components/StencilReactProps.ts
@@ -1,4 +1,0 @@
-export interface StencilReactProps {
-  class?: string;
-  style?: { [key: string]: any };
-}