+++ libs/typeorm-plus/src/lib/interface.ts
@@ -1,0 +1,3 @@
+import { ModelType, Model } from 'sequelize';
+
+export type SequelizeRepository<M = Model> = (new () => M) & ModelType;