+++ libs/stencil-toolkit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.10 (2020-01-01)
+
+**Note:** Version bump only for package @aiao/stencil-toolkit
+
+
+
+
+
 ## 0.0.9 (2019-12-05)
 
 **Note:** Version bump only for package @aiao/stencil-toolkit