+++ libs/util/CHANGELOG.md
@@ -3,4 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.3.0 (2020-01-01)
+
+
+### Features
+
+* **libs:** add angular lazy module libs ([cf1f5ec](https://github.com/aiao-io/aiao/commit/cf1f5ec71dc2213cb7edd6622a43b5ff835bf139))
+
+
+
+
+
 # 0.2.0 (2019-12-04)